+++ core/include/bertini2/endgames/base_endgame.hpp
@@ -43,7 +43,7 @@
 #include "bertini2/mpfr_complex.hpp"
 #include "bertini2/limbo.hpp"
 
-#include "bertini2/system.hpp"
+#include "bertini2/systems/system.hpp"
 
 #include "bertini2/enable_permuted_arguments.hpp"
